refactor(admin): dedupe format helpers, merge entitlement dialog, group team routes
Post-merge quality cleanup from the #426 review:

- Extract getInitials / formatDate / formatStorageUsage into src/lib/format.ts
  and migrate the admin users + teams pages off their local copies (also fixes
  a latent crash: users pages used part[0].toUpperCase() on empty segments).
  Unlimited quota now renders as ∞ consistently.
- Merge GrantUserEntitlementDialog + GrantOrgEntitlementDialog into one
  GrantEntitlementDialog keyed by a { kind: 'user' | 'team' } target; the
  dialog's labels move to a shared admin.entitlement.* namespace and the
  per-namespace dialog-only keys are removed.
- Move the org entitlement CRUD endpoints from /api/admin/quotas/:orgId to
  /api/admin/teams/:orgId so all team-admin functionality lives under one
  route tree; /api/admin/quotas keeps only the overview list. Frontend wrappers
  and tests follow.

No behavior change. lint/typecheck/test/test:cf all green.
