+++ src/i18n/locales/en.json
@@ -390,8 +390,6 @@
   "admin.users.disabled": "Disabled",
   "admin.users.enable": "Enable",
   "admin.users.disable": "Disable",
-  "admin.users.quotaUnit": "Quota unit",
-  "admin.users.positiveQuotaRequired": "Quota must be greater than 0",
   "admin.users.manageEntitlements": "Manage entitlements",
   "admin.users.entitlementsFor": "Quota entitlements for {{name}}",
   "admin.users.entitlementsUsage": "Current usage: {{used}} / {{total}}",
@@ -399,7 +397,6 @@
   "admin.users.entitlementAmount": "Amount",
   "admin.users.entitlementExpires": "Expires",
   "admin.users.entitlementStatus": "Status",
-  "admin.users.entitlementNote": "Note",
   "admin.users.entitlementSource": "Source",
   "admin.users.entitlementTypePlan": "Plan",
   "admin.users.entitlementTypeGrant": "Grant",
@@ -408,17 +405,9 @@
   "admin.users.entitlementSourceOrder": "Order",
   "admin.users.noExpiry": "No expiry",
   "admin.users.noEntitlements": "No entitlements yet.",
-  "admin.users.grantEntitlement": "Grant entitlement",
   "admin.users.addEntitlement": "Add entitlement",
-  "admin.users.grantEntitlementFor": "Grant entitlement to {{name}}",
-  "admin.users.grantEntitlementDescription": "Add a storage or traffic entitlement to this user's personal organization.",
-  "admin.users.entitlementGranted": "Entitlement granted",
   "admin.users.entitlementActions": "Actions",
   "admin.users.editEntitlement": "Edit",
-  "admin.users.editEntitlementFor": "Edit entitlement for {{name}}",
-  "admin.users.editEntitlementDescription": "Update this admin-granted entitlement. Other sources are managed automatically and cannot be edited here.",
-  "admin.users.saveEntitlement": "Save changes",
-  "admin.users.entitlementUpdated": "Entitlement updated",
   "admin.users.revokeEntitlement": "Revoke",
   "admin.users.revokeEntitlementTitle": "Revoke entitlement",
   "admin.users.revokeEntitlementConfirm": "Revoke this entitlement for {{name}}? Their available quota will be reduced immediately.",
@@ -1704,15 +1693,17 @@
   "admin.teams.entitlementRevoked": "Entitlement revoked",
   "admin.teams.active": "Active",
   "admin.teams.revoked": "Revoked",
-  "admin.teams.positiveQuotaRequired": "Amount must be a positive number",
-  "admin.teams.entitlementUpdated": "Entitlement updated",
-  "admin.teams.entitlementGranted": "Quota granted",
-  "admin.teams.editEntitlementFor": "Edit entitlement of {{name}}",
-  "admin.teams.grantEntitlementFor": "Grant quota to {{name}}",
-  "admin.teams.editEntitlementDescription": "Adjust the granted capacity, expiry, or note.",
-  "admin.teams.grantEntitlementDescription": "Grant additional storage capacity to this team.",
-  "admin.teams.quotaUnit": "Unit",
-  "admin.teams.entitlementNote": "Note (optional)",
-  "admin.teams.saveEntitlement": "Save",
-  "admin.teams.grantEntitlement": "Grant quota"
+  "admin.entitlement.grantFor": "Grant quota to {{name}}",
+  "admin.entitlement.editFor": "Edit entitlement of {{name}}",
+  "admin.entitlement.grantDescription": "Grant additional storage capacity.",
+  "admin.entitlement.editDescription": "Adjust the granted capacity, expiry, or note.",
+  "admin.entitlement.amount": "Amount",
+  "admin.entitlement.unit": "Unit",
+  "admin.entitlement.expires": "Expires (optional)",
+  "admin.entitlement.note": "Note (optional)",
+  "admin.entitlement.grant": "Grant quota",
+  "admin.entitlement.save": "Save",
+  "admin.entitlement.granted": "Quota granted",
+  "admin.entitlement.updated": "Entitlement updated",
+  "admin.entitlement.positiveQuotaRequired": "Amount must be a positive number"
 }
