+++ src/i18n/locales/zh.json
@@ -390,8 +390,6 @@
   "admin.users.disabled": "已禁用",
   "admin.users.enable": "启用",
   "admin.users.disable": "禁用",
-  "admin.users.quotaUnit": "配额单位",
-  "admin.users.positiveQuotaRequired": "配额必须大于 0",
   "admin.users.manageEntitlements": "管理权益",
   "admin.users.entitlementsFor": "{{name}} 的额度权益",
   "admin.users.entitlementsUsage": "当前用量：{{used}} / {{total}}",
@@ -399,7 +397,6 @@
   "admin.users.entitlementAmount": "额度",
   "admin.users.entitlementExpires": "过期时间",
   "admin.users.entitlementStatus": "状态",
-  "admin.users.entitlementNote": "备注",
   "admin.users.entitlementSource": "来源",
   "admin.users.entitlementTypePlan": "套餐",
   "admin.users.entitlementTypeGrant": "赠送",
@@ -408,17 +405,9 @@
   "admin.users.entitlementSourceOrder": "订单",
   "admin.users.noExpiry": "长期有效",
   "admin.users.noEntitlements": "暂无额度权益。",
-  "admin.users.grantEntitlement": "发放权益",
   "admin.users.addEntitlement": "新增权益",
-  "admin.users.grantEntitlementFor": "给 {{name}} 发放权益",
-  "admin.users.grantEntitlementDescription": "给该用户的个人组织增加一条存储或流量权益。",
-  "admin.users.entitlementGranted": "权益已发放",
   "admin.users.entitlementActions": "操作",
   "admin.users.editEntitlement": "编辑",
-  "admin.users.editEntitlementFor": "编辑 {{name}} 的权益",
-  "admin.users.editEntitlementDescription": "修改这条后台发放的权益。其他来源的权益由系统自动管理，无法在此编辑。",
-  "admin.users.saveEntitlement": "保存修改",
-  "admin.users.entitlementUpdated": "权益已更新",
   "admin.users.revokeEntitlement": "撤销",
   "admin.users.revokeEntitlementTitle": "撤销权益",
   "admin.users.revokeEntitlementConfirm": "撤销 {{name}} 的这条权益？其可用额度将立即减少。",
@@ -1704,15 +1693,17 @@
   "admin.teams.entitlementRevoked": "配额授予已撤销",
   "admin.teams.active": "生效中",
   "admin.teams.revoked": "已撤销",
-  "admin.teams.positiveQuotaRequired": "容量必须为正数",
-  "admin.teams.entitlementUpdated": "配额授予已更新",
-  "admin.teams.entitlementGranted": "配额已授予",
-  "admin.teams.editEntitlementFor": "编辑 {{name}} 的配额授予",
-  "admin.teams.grantEntitlementFor": "为 {{name}} 授予配额",
-  "admin.teams.editEntitlementDescription": "调整授予的容量、过期时间或备注。",
-  "admin.teams.grantEntitlementDescription": "为该团队授予额外的存储容量。",
-  "admin.teams.quotaUnit": "单位",
-  "admin.teams.entitlementNote": "备注（可选）",
-  "admin.teams.saveEntitlement": "保存",
-  "admin.teams.grantEntitlement": "授予配额"
+  "admin.entitlement.grantFor": "为 {{name}} 授予配额",
+  "admin.entitlement.editFor": "编辑 {{name}} 的配额授予",
+  "admin.entitlement.grantDescription": "授予额外的存储容量。",
+  "admin.entitlement.editDescription": "调整授予的容量、过期时间或备注。",
+  "admin.entitlement.amount": "容量",
+  "admin.entitlement.unit": "单位",
+  "admin.entitlement.expires": "过期时间（可选）",
+  "admin.entitlement.note": "备注（可选）",
+  "admin.entitlement.grant": "授予配额",
+  "admin.entitlement.save": "保存",
+  "admin.entitlement.granted": "配额已授予",
+  "admin.entitlement.updated": "配额授予已更新",
+  "admin.entitlement.positiveQuotaRequired": "容量必须为正数"
 }
