+++ src/lib/format.ts
@@ -1,6 +1,21 @@
-export function formatDate(timestamp: string): string {
-  const d = new Date(timestamp)
+export function formatDate(value: number | string): string {
+  const d = new Date(value)
   return Number.isNaN(d.getTime()) ? '—' : d.toLocaleDateString()
+}
+
+// Up to two uppercase initials from a display name; safe on empty/odd input.
+export function getInitials(name: string): string {
+  return name
+    .split(/\s+/)
+    .filter(Boolean)
+    .slice(0, 2)
+    .map((part) => part[0]?.toUpperCase() ?? '')
+    .join('')
+}
+
+// "used / total" with ∞ for an unlimited (0 or negative) quota.
+export function formatStorageUsage(used: number, total: number): string {
+  return `${formatSize(used)} / ${total <= 0 ? '∞' : formatSize(total)}`
 }
 
 export function formatSize(bytes: number): string {
